Fix: change decimal float

+++ db/migrate/20170227161039_create_users.rb
@@ -4,7 +4,7 @@ class CreateUsers < ActiveRecord::Migration[5.0]
       t.string :name
       t.string :last_name
       t.string :email
-      t.decimal :qualification
+      t.float :qualification
       t.belongs_to :city, index: true
 
       t.timestamps
